backend/pyproject.toml
[project]
name = "accops-backend"
version = "0.1.0"
description = "AccOps - Backend API"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "cryptography>=46.0.0",
    "drissionpage>=4.1.1.2",
    "httpx>=0.28.0",
    "pyotp>=2.9.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.9",
    "fastapi>=0.135.1",
    "uvicorn>=0.41.0",
    "python-jose[cryptography]>=3.5.0",
    "python-multipart>=0.0.22",
    "bcrypt>=5.0.0",
    "websockets>=16.0",
    "alembic>=1.18.4",
]

[[tool.uv.index]]
url = "https://pypi.org/simple/"
default = true

[dependency-groups]
dev = [
    "pytest>=9.0.3",
    "pytest-cov>=5.0",
    "ruff>=0.7",
    "mypy>=1.13",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "B", "UP", "SIM"]
# B008: FastAPI Depends() 模式; SIM105/108: 风格偏好; B904/B023: 后续手工修
ignore = ["E501", "B008", "SIM105", "SIM108", "B904", "B023", "E701", "SIM102", "B007"]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["E402"]
"alembic/versions/*" = ["E402", "F401"]

[tool.mypy]
python_version = "3.11"
ignore_missing_imports = true
disallow_untyped_defs = false
warn_unused_ignores = true
